requirements.txt
asgiref==3.8.1
contourpy==1.3.0
cycler==0.12.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fonttools==4.53.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pillow==11.0.0
plotly==5.24.0
pygame==2.3.0
PyJWT==2.10.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
tenacity==9.0.0
